estructura de funciones y login de shelter

+++ backend/src/controller/pet.controller.js
@@ -1,26 +1,30 @@
 import petService from "../services/pet.service.js";
 import fs from "fs-extra";
-import {uploadImage, deleteImage} from "../helpers/cloudinary.js"; 
+import { uploadImage, deleteImage } from "../helpers/cloudinary.js";
 
 const petController = {
   createPet: async (req, res) => {
     try {
       const data = req.body;
-      if(!Array.isArray(req.files?.image)){
-        req.files.image = [req.files.image];
-      }
-      if (req.files?.image){
-        const imagesToUpload = [];
-        for (const image of req.files.image) {
-          const result = await uploadImage(image.tempFilePath); 
-          fs.unlink(image.tempFilePath)
-          const folder = result.public_id; 
-          const url = result.secure_url; 
-          const imgs = {folder, url};
-          imagesToUpload.push(imgs); 
+      if (req.files && req.files.image) {
+        if (!Array.isArray(req.files?.image)) {
+          req.files.image = [req.files.image];
         }
-        data.images = imagesToUpload; 
-    }
+        if (req.files?.image) {
+          const imagesToUpload = [];
+          for (const image of req.files.image) {
+            const result = await uploadImage(image.tempFilePath);
+            fs.unlink(image.tempFilePath);
+            const folder = result.public_id;
+            const url = result.secure_url;
+            const imgs = { folder, url };
+            imagesToUpload.push(imgs);
+          }
+          data.images = imagesToUpload;
+          console.log(imagesToUpload);
+        }
+      }
+
       const pet = await petService.createPet(data);
       return res.status(201).json(pet);
     } catch (error) {
